refactor: PictureService와 ContentService 통합
PictureService에 있던 로직은 Content에 사진을 추가하는 로직이고
그 양이 많지 않아서, ContentService에 통합
