+++ src/main/java/com/strcat/service/ContentService.java
@@ -6,16 +6,19 @@ import com.strcat.dto.CreateContentReqDto;
 import com.strcat.exception.NotAcceptableException;
 import com.strcat.repository.BoardRepository;
 import com.strcat.repository.ContentRepository;
+import com.strcat.repository.PictureRepository;
 import com.strcat.util.SecureDataUtils;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 @Service
 @RequiredArgsConstructor
 public class ContentService {
     private final ContentRepository contentRepository;
     private final BoardRepository boardRepository;
+    private final PictureRepository pictureRepository;
     private final SecureDataUtils secureDataUtils;
 
     public Content create(CreateContentReqDto dto, String encryptedBoardId) {
@@ -27,5 +30,20 @@ public class ContentService {
         }
         Board board = optionalBoard.get();
         return contentRepository.save(new Content(dto.getWriter(), dto.getText(), dto.getPhotoUrl(), board));
+    }
+
+    public String postPicture(String encryptedBoardId, MultipartFile picture) {
+        if (isInvalidContentType(picture.getContentType())) {
+            throw new NotAcceptableException("처리할 수 없는 파일 형식입니다.");
+        }
+
+        return pictureRepository.postPicture(encryptedBoardId, picture);
+    }
+
+    private boolean isInvalidContentType(String contentType) {
+        return switch (contentType) {
+            case "image/jpeg", "image/png", "image/jpg" -> false;
+            default -> true;
+        };
     }
 }
